extract points client functions
